requirements.txt
Flask==2.3.2
flask-cors==3.1.0
python-dateutil==2.8.2
Babel==2.12.1
